ci: run public_repo_test in separate job
This PR separates out the public_repo tests from other unit/e2e tests
so those tests can return faster
