+++ .github/workflows/tests.yml
@@ -130,6 +130,51 @@ jobs:
           export PATH=/github/home/.local/bin:$PATH
           pip3 install pipenv
           pipenv install --dev
+      - name: Test semgrep
+        run: |
+          cd semgrep
+          export PATH=/github/home/.local/bin:$PATH
+          pipenv run pytest -n auto -v --tb=short --durations=0 tests/
+      - name: Generate test matrix data
+        run: |
+          cd semgrep
+          pipenv run ../scripts/generate_test_matrix.py --json --directory ../semgrep-core/tests/ --output-file test-matrix.json
+      - name: Upload artifacts
+        uses: actions/upload-artifact@v1
+        with:
+          name: test-matrix
+          path: semgrep/test-matrix.json
+
+  public-repos-test:
+    name: run semgrep on public repos
+    runs-on: ubuntu-latest
+    needs: [build-core]
+    strategy:
+      matrix:
+        python: [3.7]
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+        with:
+          submodules: true
+      - name: Setup Python
+        uses: actions/setup-python@v1
+        with:
+          python-version: ${{ matrix.python }}
+      - name: Download artifacts
+        uses: actions/download-artifact@v1
+        with:
+          name: ocaml-build-artifacts
+      - name: Install artifacts
+        run: |
+          tar xf ocaml-build-artifacts/ocaml-build-artifacts.tgz
+          sudo cp ocaml-build-artifacts/bin/* /usr/bin
+      - name: Install semgrep
+        run: |
+          cd semgrep
+          export PATH=/github/home/.local/bin:$PATH
+          pip3 install pipenv
+          pipenv install --dev
       - uses: actions/cache@v2
         with:
           path: ~/.cache/semgrep-cache
@@ -141,18 +186,9 @@ jobs:
         run: |
           cd semgrep
           export PATH=/github/home/.local/bin:$PATH
-          make ci-test
+          pipenv run pytest -n auto -v --tb=short --durations=0 tests/qa/run_public_repos.py
         env:
           GITHUB_REPO_CACHE: ~/.cache/semgrep-cache
-      - name: Generate test matrix data
-        run: |
-          cd semgrep
-          pipenv run ../scripts/generate_test_matrix.py --json --directory ../semgrep-core/tests/ --output-file test-matrix.json
-      - name: Upload artifacts
-        uses: actions/upload-artifact@v1
-        with:
-          name: test-matrix
-          path: semgrep/test-matrix.json
 
   # Run each benchmark twice to decrease effect of natural variance
   benchmark-test:
